# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Offers API",
    "version": "1.0.0",
    "description": "API for fetching different types of offers (vehicle, real estate, and commercial)."
  },
  "paths": {
    "/auth": {
      "post": {
        "summary": "Login user or company",
        "description": "Authenticates the user or company and returns an access token and refresh token.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Login successful"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized (Incorrect password)"
          },
          "404": {
            "description": "Not Found (Email not found)"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}